Get deals at a specific location using Groupon API and Google Maps API
Find deals for the user's location using Groupon API and Google
locations and Google Maps API. Then a list view and a map view was
provided for all the deals on two separate tabs. The user could also
change the location using a drop down menu. Again here a wishlist was
provided to store user specific deals. A login page was created to enter
the App and the user was authenticated based on his credentials stored
in Parse.com